+++ website/src/routes/(docs)/methods/api/(types)/GetFieldInputConfig/properties.ts
@@ -41,4 +41,11 @@ export const properties: Record<string, PropertyProps> = {
       ],
     },
   },
+  dirtyOnly: {
+    type: 'boolean',
+    default: {
+      type: 'boolean',
+      value: false,
+    },
+  },
 };